refactor(R2.2): 配置更新分发 790 行摘出 ReadBookViewModel
剩余最大一块。ReadBookViewModel 4903 → 4127 行。

handleConfigUpdate（671 行 158 分支）+ toReadStyleMutation + 5 个 mutation
辅助共 790 行摘成 ReadConfigUpdateDelegate。

和前三个域不同，这块**没有自持状态**——它写的 menuConfig 是阅读器 chrome
的状态、被菜单栏读，所以留在 ReadBookUiState，经 Host 回写。依赖面本就
很窄（79× readSettingsRepository / 77× viewModelScope / 39× _uiState /
4× _effects / 1× styleGateway，无 DAO、无 ReadBook、无 context），
Host 只需 4 个成员：
- menuConfig（2 处分支要读旧值做回退判断）
- updateMenuConfig(transform)（34 处单行 + 2 处多行写入）
- refreshConfigSnapshots()（styleMutation == null 的那条分支，是排版快照
  三条重建通道里的第三条，R1.1 的注释仍在 collectReadStyle 上）
- resetDayNightReminderDismissal()（AutoSuggestDayNight 打开时清标记，
  标记本身属 reminder 域，留在 VM）

5 个 xxxMutation 辅助提成同包顶层 internal——VM 里另有 10 处在用
（selectFont / colorSelected 等），不能一起搬走。

改写用括号配平的结构化转换器做，不用正则。搬迁后做了对括号不敏感的归一化
多重集比对：**524 条语义行完全对上，零漏搬零新增**。

护栏：DOMAINS 表加一条。该域无状态，stateFields 为空，靠
stateTypes = listOf("is ConfigUpdate.") 守住「158 分支不回流 VM」。
已注入回归反证两条断言变红。

验证：testAppDebugUnitTest(426 passed) + lintAppDebug +
verifyConfigArchitecture + assembleAppDebug 全绿（与 CI verify.yml 同一组）。
